Group similar files into groups
