+++ Sources/PhotoCaptureViewController/CaptureManager.swift
