+++ Sources/PhotoCaptureViewController/ImagePickerAdapter.swift
