+++ Sources/PhotoCaptureViewController/PhotoCollectionViewCell.swift
