+++ Sources/PhotoCaptureViewController/PhotoCollectionViewLayout.swift
